Add a proxy to allow secondary assets to load
The Vite dev server rewrites asset includes (css url() calls, for
instance) but only does so at the path level regardless of whether a
full URL with port has been specified as the base in Vite's config. A
proxy is how vite_ruby handles this. Ours is incomplete but seems to
work well enough for our purposes.